requirements.txt
astroid==1.6.4
certifi==2018.4.16
chardet==3.0.4
idna==2.6
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
numpy==1.14.3
opencv-python==3.4.1.15
Pillow==5.0.0
pylint==1.9.1
pytesseract==0.2.2
requests==2.18.4
selenium==3.12.0
six==1.11.0
urllib3==1.22
wrapt==1.10.11
